remove ram comment about Parallel::Downloader
Using the `build_response` one can pass in a sub that is intended to build for each request the response emitted at the end of the batch. The default build process involves storing the request and response in an array, but can also store the response in a file and return a filename instead of the response, or simply return nothing at all.

The only real downside is that it's a one-shot tool.